+++ bores/src/main/java/com/tgfcodes/bores/service/ProdutoService.java
@@ -5,24 +5,34 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.tgfcodes.bores.exception.NegocioException;
 import com.tgfcodes.bores.model.Produto;
 import com.tgfcodes.bores.repository.ProdutoRepository;
+import com.tgfcodes.bores.repository.SubcategoriaRepository;
 
 @Service
 public class ProdutoService {
 
 	@Autowired
 	private ProdutoRepository produtoRepository;
+	@Autowired
+	private SubcategoriaRepository subcategoriaRepository;
+	//@Autowired
+	//private PedidoRepository pedidoRepository;
 
 	@Transactional(readOnly = false)
 	public void salvar(Produto produto) {
+		produto.setSubcategoria((produto.getId() == null) ? this.subcategoriaRepository.getOne(produto.getSubcategoria().getId()) : produto.getSubcategoria());
 		this.produtoRepository.save(produto);
 	}
 	
 	@Transactional(readOnly = false)
 	public void excluir(Produto produto) {
+		boolean remover = false;
+		if(remover) {
+			throw new NegocioException("Produto: Não pode ser excluído. Possue pedidos associados.");
+		}
 		produto = this.buscarPorId(produto.getId());
-		//TODO: não pode excluir se pertencer a algum pedido
 		this.produtoRepository.delete(produto);
 	}
