+++ bores/src/main/java/com/tgfcodes/bores/service/SubcategoriaService.java
@@ -11,6 +11,7 @@ import com.tgfcodes.bores.exception.NegocioException;
 import com.tgfcodes.bores.model.Categoria;
 import com.tgfcodes.bores.model.Subcategoria;
 import com.tgfcodes.bores.repository.CategoriaRepository;
+import com.tgfcodes.bores.repository.ProdutoRepository;
 import com.tgfcodes.bores.repository.SubcategoriaRepository;
 
 @Service
@@ -19,38 +20,40 @@ public class SubcategoriaService {
 	@Autowired
 	private SubcategoriaRepository subcategoriaRepository;
 	@Autowired
-	private CategoriaRepository categoriaRepository; 
-	
+	private CategoriaRepository categoriaRepository;
+	@Autowired
+	private ProdutoRepository produtoRepository;
+
 	@Transactional(readOnly = false)
 	public void salvar(Subcategoria subcategoria) {
+		subcategoria.setCategoria((subcategoria.getId() == null) ? this.categoriaRepository.getOne(subcategoria.getCategoria().getId())	: subcategoria.getCategoria());
 		this.subcategoriaRepository.save(subcategoria);
 	}
-	
+
 	@Transactional(readOnly = false)
 	public void excluir(Subcategoria subcategoria) {
-		subcategoria = subcategoriaRepository.findById(subcategoria.getId()).get();
-		/*
-		 * if(!subcategoria.getProdutos().isEmpty()) { throw new
-		 * NegocioException("Subcategoria: Não pode ser excluída. Possue produtos associados."
-		 * ); }
-		 */
+		boolean remover = this.produtoRepository.existsBySubcategoria(subcategoria);
+		if (remover) {
+			throw new NegocioException("Subcategoria: Não pode ser excluída. Possue produtos associados.");
+		}
+		subcategoria = this.buscarPorId(subcategoria.getId());
 		this.subcategoriaRepository.delete(subcategoria);
 	}
-	
+
 	@Transactional(readOnly = true)
-	public List<Subcategoria> listar(Long categoriaId){
+	public List<Subcategoria> listar(Long categoriaId) {
 		Categoria categoria = this.categoriaRepository.findById(categoriaId).get();
 		return this.subcategoriaRepository.findByCategoria(categoria);
 	}
-	
+
 	@Transactional(readOnly = true)
-	public LazyDataModel<Subcategoria> pesquisar(){
+	public LazyDataModel<Subcategoria> pesquisar() {
 		return this.subcategoriaRepository.pesquisar();
 	}
-	
+
 	@Transactional(readOnly = true)
 	public Subcategoria buscarPorId(Long subcategoriaId) {
 		return this.subcategoriaRepository.findById(subcategoriaId).get();
-	} 
-	
+	}
+
 }
