+++ bores/src/main/java/com/tgfcodes/bores/service/UsuarioService.java
@@ -6,11 +6,13 @@ import org.primefaces.model.LazyDataModel;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.StringUtils;
 
 import com.tgfcodes.bores.exception.NegocioException;
 import com.tgfcodes.bores.model.Grupo;
 import com.tgfcodes.bores.model.Usuario;
 import com.tgfcodes.bores.repository.GrupoRepository;
+import com.tgfcodes.bores.repository.PedidoRepository;
 import com.tgfcodes.bores.repository.UsuarioRepository;
 
 @Service
@@ -20,19 +22,37 @@ public class UsuarioService {
 	private UsuarioRepository usuarioRepository;
 	@Autowired
 	private GrupoRepository grupoRepository;
+	@Autowired
+	private PedidoRepository pedidoRepository;
+	//@Autowired
+	//private PasswordEncoder passwordEncoder;
+	
 	
 	@Transactional(readOnly = false)
 	public void salvar(Usuario usuario) {
-		if(usuario.getId() == null) {
+		Usuario usuarioExistente = this.usuarioRepository.findByEmailIgnoreCase(usuario.getEmail());
+		if(usuarioExistente != null && !usuario.equals(usuarioExistente)) {
 			throw new NegocioException("Email: Já esta cadastrado para outro usuário.");
+		}
+		if(usuario.getId() == null && !StringUtils.hasText(usuario.getSenha())) {
+			throw new NegocioException("Senha: Obrigatória a novos usuário.");
+		}
+		if (usuario.getId() == null || !StringUtils.hasText(usuario.getSenha())) {
+			//TODO: descomentar ao colocar segurança
+			//usuario.setSenha(this.passwordEncoder.encode(usuario.getSenha()));
+		} else if (StringUtils.hasText(usuario.getSenha())) {
+			usuario.setSenha(usuarioExistente.getSenha());
 		}
 		this.usuarioRepository.save(usuario);
 	}
 	
 	@Transactional(readOnly = false)
 	public void excluir(Usuario usuario) {
+		boolean remover = this.pedidoRepository.existsByUsuario(usuario);
+		if(remover) {
+			throw new NegocioException("Usuário: Não pode ser excluído. Possue pedidos associados.");
+		}
 		usuario = this.buscarPorId(usuario.getId());
-		//TODO: não pode excluir se pertencer a algum pedido
 		this.usuarioRepository.delete(usuario);
 	}
 	
@@ -46,7 +66,7 @@ public class UsuarioService {
 	@Transactional(readOnly = true)
 	public List<Usuario> buscarPorGrupo(Boolean status, String nameGrupo){
 		Grupo grupo = this.grupoRepository.findByNomeIgnoreCase(nameGrupo);
-		return this.usuarioRepository.buscarPorGrupo(status, grupo.getId());
+		return this.usuarioRepository.porGrupo(status, grupo.getId());
 	}
 	
 	@Transactional(readOnly = true)
